Tasj

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,9 +7,9 @@
 int main(void)
 {
 	long int sum = 0;
-	int first = 0, second = 1, i;
+	long int first = 0, second = 1, i;
 
-	printf("%d, ", first);
+	printf("%li, ", first);
 	for (i = 0; i <= 50; i++)
 	{
 		sum = first + second;
